# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  github_com_evolution-foundation_evolution-go_pkg_message_service.MarkReadStruct:
    properties:
      id:
        items:
          type: string
        type: array
      number:
        type: string
    type: object
...
info:
  contact: {}
  description: Evolution GO - whatsmeow
  title: Evolution GO
  version: "1.0"
paths:
...
  /message/markread:
    post:
      consumes:
      - application/json
      description: Mark a message as read
      parameters:
      - description: Mark a message as read
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_evolution-foundation_evolution-go_pkg_message_service.MarkReadStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Mark a message as read
      tags:
      - Message
...
  /sender/ws:
    get:
      description: Same-origin WebSocket. Token is the instance key; query instanceId
        is rejected. Access logs must redact query tokens.
      parameters:
      - description: Instance API key
        in: query
        name: token
        required: true
        type: string
      responses:
        "101":
          description: WebSocket upgrade
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/gin.H'
      summary: Sender instance event stream
      tags:
      - Sender
...
swagger: "2.0"